# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Smart Airport Ride Pooling System API
  version: 1.0.0
  description: A comprehensive API for airport ride pooling with real-time matching and dynamic pricing
  contact:
    name: API Support
    email: support@airportpooling.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
paths:
...
  /api/admin/pools:
    get:
      tags:
        - Admin
      summary: List all active pools
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: airportCode
          in: query
          schema:
            type: string
            enum: [BOM, DEL, BLR, HYD, MAA]
        - name: status
          in: query
          schema:
            type: string
            enum: [forming, locked, assigned]
      responses:
        '200':
          description: Active pools retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      pools:
                        type: array
                        items:
                          $ref: '#/components/schemas/Pool'
                      pagination:
                        type: object
        '401':
          description: Authentication required
        '403':
          description: Admin access required
  /api/admin/analytics:
    get:
      tags:
        - Admin
      summary: Get KPIs and analytics
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [1h, 24h, 7d, 30d]
            default: 24h
      responses:
        '200':
          description: Analytics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      kpis:
                        type: object
                      charts:
                        type: object
        '401':
          description: Authentication required
        '403':
          description: Admin access required
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Pool:
      type: object
      properties:
        _id:
          type: string
        airportCode:
          type: string
        rides:
          type: array
          items:
            type: string
        driverId:
          type: string
        status:
          type: string
          enum: [forming, locked, assigned, completed]
        totalSeats:
          type: integer
        totalLuggage:
          type: integer
        routeOrder:
          type: array
          items:
            type: string
        totalDeviation:
          type: integer
        createdAt:
          type: string
          format: date-time
...